Move tests for admin views to a specific test file
